pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "mcp[cli]>=1.11.0",
    "nest-asyncio>=1.6.0",
    "openai>=1.95.1",
    "python-dotenv>=1.1.1",
]
